requirements.txt
google-generativeai
PyYAML
